chore: :fire: updated outdated test suites and restructuring
