+++ src/features/dashboard/components/Table/copy-text-cell.scss
