+++ src/features/dashboard/components/Table/scopes-cell.module.scss
