+++ src/features/dashboard/components/Table/scopes-cell.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './scopes.cell.module.scss';
+import styles from './scopes-cell.module.scss';
 
 type TScopesCellProps = {
   cell: {
