+++ src/features/dashboard/components/api-token-table/index.ts
@@ -1,0 +1,3 @@
+import ApiTokenTable from './api-token-table';
+
+export default ApiTokenTable;
