+++ src/features/dashboard/components/dialogs/token-creation-dialog-success/__tests__/token-creation-dialog-success.test.tsx
@@ -39,7 +39,7 @@ describe.skip('Token Creation Dialog', () => {
       setLatestToken: setLatestTokenMock,
     }));
 
-    render(<TokenCreationDialogSuccess setToggleModal={setToggleModalMock} />);
+    render(<TokenCreationDialogSuccess setToggleModal={setToggleModalMock} is_toggle={false} />);
   });
 
   afterEach(() => {
